decorations: parse at the backend and serve assets immutable
